+++ feature-libs/cart/quick-order/components/container/index.ts
@@ -1,0 +1,2 @@
+export * from './quick-order-container.component';
+export * from './quick-order-container.module';
