+++ feature-libs/cart/quick-order/components/container/quick-order-container.component.ts
@@ -1,8 +1,8 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
 
 @Component({
-  selector: 'cx-quick-order-wrapper',
-  templateUrl: './quick-order-wrapper.component.html',
+  selector: 'cx-quick-order-container',
+  templateUrl: './quick-order-container.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class QuickOrderComponent {}
