+++ feature-libs/cart/quick-order/components/public_api.ts
@@ -1,2 +1,2 @@
-export * from './wrapper/index';
+export * from './container/index';
 export * from './quick-order-components.module';
