--- feature-libs/cart/quick-order/components/wrapper/index.ts
@@ -1,2 +1,0 @@
-export * from './quick-order-wrapper.component';
-export * from './quick-order-wrapper.module';
