--- feature-libs/cart/quick-order/components/wrapper/quick-order-wrapper.component.html
@@ -1,1 +1,0 @@
-Quick Order Wrapper Component
